Add README.md for Oasis Web Development Internship projects
This commit adds the README.md file to the repository, providing an overview of the projects completed during my 1-month Oasis Web Development Internship. The file includes details about the technologies used, such as HTML, CSS, JavaScript, as well as an introduction to the skills I gained in frontend development. The README also provides a link to my LinkedIn profile for further connection and collaboration.